[C2-08] move leaderboard to component folder + remove profile img border
